pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawfs"
version = "0.2.0"
description = "Content-addressed file system with refcounted GC"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "sqlmodel>=0.0.16",
    "python-multipart>=0.0.9",
    "click>=8.1",
]

[project.optional-dependencies]
azure = ["azure-storage-blob>=12.19"]
test = ["pytest>=8.0", "httpx>=0.27"]
dev = ["ruff>=0.4", "pytest>=8.0", "httpx>=0.27"]

[project.scripts]
clawfs = "clawfs.cli:cli"

[tool.setuptools.packages.find]
include = ["clawfs*"]
